double Price округление цен

+++ DomesticTransport/Model/TruckRate.cs
@@ -6,9 +6,9 @@
         public string PlaceDelivery { get; set; }
         public string City { get; set; }
         public string Company { get; set; }
-        public int PriceFirstPoint { get; set; }
-        public int PriceAddPoint { get; set; }
+        public double PriceFirstPoint { get; set; }
+        public double PriceAddPoint { get; set; }
         public double Tonnage { get; set; }
-        public int TotalDeliveryCost { get; set; }
+        public double TotalDeliveryCost { get; set; }
     }
 }
